[ENH] Consolidating physio api
